Using App-Message Contents
to get JSON objects and pass them to appropriate process methods of Controllers. The process methods will get search criteria from the JSON objects if possible and return Products.

Modified signature of abstract method Controller.process to take JSONObject valuesPassedToController ( e.g., {"ingredients": ["cream", "sugar"]} ).

Modified ExtendedController.process to take the same input and, in the main body, not only get search criteria and get products matching search criteria, but also return empty Products if the object that is supposed to have an ingredients list doesn't or has an empty ingredients list.

Added to IceCreamApplicationMessage a method to get a String representing the body of a client message.

Implemented SearchCriteria.

Added to MessageHandler a method to add controllers. Rewrote handle to split the body of a client message into (key_plus_values)'s. Each key_plus_values is split into a key and values, which forms a row in a hashMap, if it is valid and hasn't already been added to the hashMap. For each row, the values are decoded and used to construct a JSON object, which is passed to the process method of a Controller identified by the key in the row.